Split receiver code. Added last entry time for host
