+++ database/rrd.h
@@ -973,6 +973,8 @@ static inline time_t rrdset_first_entry_t(RRDSET *st) {
     }
 }
 
+time_t rrdhost_last_entry_t(RRDHOST *h);
+
 // get the last slot updated in the round robin database
 #define rrdset_last_slot(st) ((size_t)(((st)->current_entry == 0) ? (st)->entries - 1 : (st)->current_entry - 1))
